+++ CHANGES.md
@@ -8,3 +8,17 @@ Recently completed and planned work is tracked here.
 - Updated `README.md`
 - Upgraded `hashicorp/aws` provider
 - Upgraded `name` module
+
+## [0.1.2](.)
+- Added `enable_eks` variable
+- Added `enable_k8_auth` variable
+- Added `kms_key_arn` where needed
+- Improved variable descriptions
+- Removed `hashicorp/null` provider
+- Resolved `tfsec` warnings
+- Updated `ec2_work` module
+- Updated `ecr` module
+- Updated `eks` module
+- Updated `k8_auth` module
+- Updated `name` module
+- Updated `sg_ingress_open` module